backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
ultralytics
pillow
google-generativeai
azure-cognitiveservices-speech